backend/requirements.txt
fastapi
uvicorn
torch
matplotlib
numpy
scikit-learn
torchvision
wheel
coincurve
redis
celery
flower
sqlalchemy
python-dotenv
psycopg2